+++ analysis_options.yaml
@@ -1,1 +1,1 @@
-include: package:very_good_analysis/analysis_options.3.0.0.yaml
+include: package:very_good_analysis/analysis_options.3.1.0.yaml
