+++ example/pubspec.yaml
@@ -4,11 +4,10 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.17.0 <3.0.0"
-  flutter: ">=3.0.0"
+  sdk: ">=2.18.2 <3.0.0"
+  flutter: ">=3.3.5"
 
 dependencies:
-  cupertino_icons: ^1.0.4
   flutter:
     sdk: flutter
   flutter_lazy_indexed_stack:
@@ -17,7 +16,7 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  very_good_analysis: ^3.0.0
+  very_good_analysis: ^3.1.0
 
 flutter:
   uses-material-design: true
