+++ pubspec.yaml
@@ -5,8 +5,8 @@ homepage: https://github.com/marcossevilla/lazy_indexed_stack
 repository: https://github.com/marcossevilla/lazy_indexed_stack
 
 environment:
-  sdk: ">=2.17.0 <3.0.0"
-  flutter: ">=3.0.0"
+  sdk: ">=2.18.2 <3.0.0"
+  flutter: ">=3.3.5"
 
 dependencies:
   flutter:
@@ -15,4 +15,4 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  very_good_analysis: ^3.0.0
+  very_good_analysis: ^3.1.0
